+++ tests/index.js
@@ -5,6 +5,7 @@
 const assert = require('assert');
 const fs = require('fs');
 const path = require('path');
+const arrayIncludes = require('array-includes');
 
 const plugin = require('..');
 
@@ -24,10 +25,29 @@ describe('all rule files should be exported by the plugin', () => {
 
 describe('rule documentation files have the correct content', () => {
   const MESSAGES = {
+    configs: '💼 This rule is enabled in the following [configs](https://github.com/jsx-eslint/eslint-plugin-react#shareable-configurations):',
+    configsOff: 'This rule is disabled in the following configs:',
     deprecated: '❌ This rule is deprecated.',
     fixable: '🔧 This rule is automatically fixable using the `--fix` [flag](https://eslint.org/docs/latest/user-guide/command-line-interface#--fix) on the command line.',
     hasSuggestions: '💡 This rule provides editor [suggestions](https://eslint.org/docs/developer-guide/working-with-rules#providing-suggestions).',
   };
+
+  function getConfigsForRule(ruleName, checkForEnabled) {
+    const configNames = [];
+    Object.keys(plugin.configs).forEach((configName) => {
+      const value = plugin.configs[configName].rules[`react/${ruleName}`];
+      const isOn = arrayIncludes([2, 'error'], value);
+      const isOff = arrayIncludes([0, 'off'], value);
+      if (value !== undefined && ((checkForEnabled && isOn) || (!checkForEnabled && isOff))) {
+        configNames.push(configName);
+      }
+    });
+    return configNames.sort();
+  }
+
+  function configNamesToList(configNames) {
+    return `\`${configNames.join('`, `')}\``;
+  }
 
   ruleFiles.forEach((ruleName) => {
     it(ruleName, () => {
@@ -45,8 +65,10 @@ describe('rule documentation files have the correct content', () => {
       const unexpectedNotices = [];
       if (rule.meta.deprecated) {
         expectedNotices.push('deprecated');
+        unexpectedNotices.push('configs');
       } else {
         unexpectedNotices.push('deprecated');
+        expectedNotices.push('configs');
       }
       if (rule.meta.fixable) {
         expectedNotices.push('fixable');
@@ -66,6 +88,15 @@ describe('rule documentation files have the correct content', () => {
         if (expectedNotice === 'deprecated' && documentLines[currentLineNumber + 1] !== MESSAGES[expectedNotice] && documentLines[currentLineNumber + 1].startsWith(MESSAGES[expectedNotice])) {
           // Allow additional rule-specific information at the end of the deprecation notice line.
           assert.ok(true, `includes ${expectedNotice} notice`);
+        } else if (expectedNotice === 'configs') {
+          // Check that the rule specifies its configs.
+          const configsOn = getConfigsForRule(ruleName, true);
+          let expectedMessage = `${MESSAGES.configs} ${configNamesToList(configsOn)}.`;
+          const configsOff = getConfigsForRule(ruleName, false);
+          if (configsOff.length > 0) {
+            expectedMessage += ` ${MESSAGES.configsOff} ${configNamesToList(configsOff)}.`;
+          }
+          assert.strictEqual(documentLines[currentLineNumber + 1], expectedMessage, 'includes configs notice');
         } else {
           // Otherwise, just check the whole line.
           assert.strictEqual(documentLines[currentLineNumber + 1], MESSAGES[expectedNotice], `includes ${expectedNotice} notice`);